Add innercontainer support.
 This is necessary for our usecase where we use ElementQueries. ResizeSensor
places a div inside the contentarea container element. In order to be able
to use css:first-child and friends, we need to contain the block contents
within a second-level wrapper which we all innercontainer. This innercontainer
is then at the same level as the sensor-div created by ResizeSensor and
we achieve what we want.